requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
pdfkit==1.0.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
matplotlib==3.9.1.post1